backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.6.1
orjson==3.9.10